First pass of timestamp funcitonality
Wrote the timestamp algorithm, which returns unix timestamp and
formatted dates in a json object. At present, everything is in the
controller, and the algorithm is quite crude (i.e., poor from an OO
point of view). The timestamp functionality needs to be refactored out
into a model (or more probably, several models).